requirements.txt
streamlit
PyMuPDF
python-docx
pandas
pillow
pytesseract


